feat: updated countCharPar to work with vector operations right until the end

+++ Aufgabe_A/measurement/string_manipulation_par.h
@@ -1,8 +1,10 @@
 #ifndef STRING_MAN_PAR_H
 #define STRING_MAN_PAR_H
 
+#include <stdint.h>
+
 void init_register();
 int toUppercasePar(char *string, int len_string);
 int toLowercasePar(char *string, int len_string);
-int countCharPar(char *string, int len_string);
+int64_t countCharPar(char *string, int len_string);
 #endif
